requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-generativeai>=0.3.0
openai>=1.0.0
gspread>=5.10.0
oauth2client>=4.1.3
